Break view builders into separate files
