+++ refbox/src/app/view_builders/mod.rs
@@ -1,0 +1,28 @@
+use super::*;
+
+pub mod keypad_pages;
+pub use keypad_pages::*;
+
+pub mod configuration;
+pub use configuration::*;
+
+pub mod confirmation;
+pub use confirmation::*;
+
+pub mod list_selector;
+pub(super) use list_selector::*;
+
+pub mod main_view;
+pub use main_view::*;
+
+pub mod penalties;
+pub use penalties::*;
+
+pub mod score_edit;
+pub use score_edit::*;
+
+pub mod shared_elements;
+pub use shared_elements::*;
+
+pub mod time_edit;
+pub use time_edit::*;
